api_gateway/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.1.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
respx>=0.21.1
